add link to vulnarabilities list to security page
🔍 chore(.opencommitignore): add ignore rules for lerna-debug.log, .DS_Store, .vscode, and .idea/
Added ignore rules for files and directories that are not relevant to the project, such as logs, diagnostic reports, runtime data, and cache directories. Also added ignore rules for environment variables files and build output directories.

🚨 test(image-card.spec.tsx): add tests for ImageCard and Target components
Added tests for the ImageCard and Target components to ensure that they render correctly and apply the correct styles and classes. The tests cover the rendering of the link, image, and children passed as props, as well as the application of the "imageCardVertical" and "disabled" classes. The Target component tests also cover the rendering of the IconExternal element and the application of the "disabledTarget" class.

🎨 style(image-card.tsx): remove alt prop and use ReactNode for image prop
Removed

🚀 chore(security): add tests for SecOpsCodeReview and SecOpsRiskThreatModel components
✨ feat(security): add href prop to SecOpsRiskThreatModel component
✨ feat(security): add Next.js Image component to SecOpsCodeReview and SecOpsRiskThreatModel components
The tests for SecOpsCodeReview and SecOpsRiskThreatModel components have been added to ensure that they render correctly. The href prop has been added to the SecOpsRiskThreatModel component to allow for more flexibility in the application. The Next.js Image component has been added to both components to improve performance and accessibility.